# docs/swagger.yaml
definitions:
  model.BaseResp:
    properties:
      code:
        description: 结果码
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.CreateLinkDefineReq:
    properties:
      comment:
        description: 字段注释
        type: string
      name:
        description: 项目名
        type: string
      paramList:
        description: 参数列表
        items:
          $ref: '#/definitions/model.LinkParamDefine'
        type: array
    type: object
  model.CreateLinkDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 创建成功后的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.CreateLinkInstanceReq:
    properties:
      comment:
        description: 注释
        type: string
      defineID:
        description: 对应 定义 id
        type: integer
      firstLinkDefineID:
        description: 第一个 link 信息
        type: integer
      firstLinkInstanceID:
        type: integer
      generateLanguage:
        description: 实现工程所用语言及版本
        type: string
      name:
        description: 工程名
        type: string
      paramList:
        description: 参数信息
        items:
          $ref: '#/definitions/model.LinkParamInstance'
        type: array
    type: object
  model.CreateLinkInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 创建成功后的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.CreateNodeDefineReq:
    properties:
      belongLinkDefineID:
        description: 所属link id
        type: integer
      comment:
        description: 注释
        type: string
      name:
        description: node名
        type: string
      paramList:
        description: 参数列表
        items:
          $ref: '#/definitions/model.NodeParamDefine'
        type: array
      preNodeDefineID:
        description: 前一个nodeID , 第一个写-1
        type: integer
    type: object
  model.CreateNodeDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 创建成功后的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.DeleteLinkDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 删除的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.DeleteLinkInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 删除的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.DeleteNodeDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 删除的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.GetLinkDefineListResp:
    properties:
      code:
        description: 结果码
        type: integer
      list:
        description: 项目列表
        items:
          $ref: '#/definitions/model.LinkDefineInfo'
        type: array
      msg:
        description: 提示信息
        type: string
      sum:
        description: 总数据条数
        type: integer
    type: object
  model.GetLinkDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      info:
        $ref: '#/definitions/model.LinkDefineInfo'
        description: 结构体信息
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
  model.GetLinkInstanceListResp:
    properties:
      code:
        description: 结果码
        type: integer
      list:
        description: 项目列表
        items:
          $ref: '#/definitions/model.LinkInstanceInfo'
        type: array
      msg:
        description: 提示信息
        type: string
      sum:
        description: 总数据条数
        type: integer
    type: object
  model.GetLinkInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      info:
        $ref: '#/definitions/model.LinkInstanceInfo'
        description: 结构体信息
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
  model.GetNodeDefineListResp:
    properties:
      code:
        description: 结果码
        type: integer
      list:
        description: node列表
        items:
          $ref: '#/definitions/model.NodeDefineInfo'
        type: array
      msg:
        description: 提示信息
        type: string
      sum:
        description: 总数据条数
        type: integer
    type: object
  model.GetNodeDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      info:
        $ref: '#/definitions/model.NodeDefineInfo'
        description: node信息
        type: object
      msg:
        description: 提示信息
        type: string
    type: object
...
  model.LinkDefineInfo:
    properties:
      comment:
        description: 注释
        type: string
      hasInstance:
        type: boolean
      id:
        type: integer
      isShared:
        description: 是否共享, 不共享的话 只能当前工程可用
        type: boolean
      name:
        description: 链名称
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.LinkParamDefine'
        type: array
      projectDefineID:
        description: 所属工程ID
        type: integer
    type: object
  model.LinkInstanceInfo:
    properties:
      comment:
        description: 注释
        type: string
      defineID:
        description: 所用 链 定义 id
        type: integer
      id:
        type: integer
      isShared:
        description: 是否共享, 不共享的话 只能当前工程可用
        type: boolean
      name:
        description: 链名称
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.LinkParamInstance'
        type: array
      projectInstanceID:
        description: 所属工程ID
        type: integer
    type: object
  model.LinkParamDefine:
    properties:
      comment:
        description: 字段注释
        type: string
      constData:
        items:
          type: integer
        type: array
      hasInstance:
        type: boolean
      id:
        type: integer
      isConst:
        description: 如果是固定值, 那么记录
        type: boolean
      isList:
        description: 是否为 列表
        type: boolean
      isParam:
        description: 位置, 1.入参 2.出参
        type: integer
      linkDefineID:
        description: 所属 链 id
        type: integer
      mapKeyParamID:
        description: 当ParamType 为 map 时
        type: integer
      mapValueParamID:
        type: integer
      name:
        description: 字段名
        type: string
      sort:
        description: 排序位置
        type: integer
      structID:
        description: 对应结构体ID, 当 FType 为 ParamTypeStruct 时 有用
        type: integer
      type:
        description: 字段类型
        type: integer
    type: object
  model.LinkParamInstance:
    properties:
      comment:
        description: 字段注释
        type: string
      defineID:
        description: 所用 链参数 定义 id
        type: integer
      id:
        type: integer
      inputConst:
        description: 如果是来自确定值, 那么记录这个值
        items:
          type: integer
        type: array
      inputType:
        description: 入参来源
        type: integer
      inputVarDefineID:
        description: 如果来自变量, 那么记录来源
        type: integer
      inputVarInstanceID:
        description: 如果来自变量, 那么记录来源
        type: integer
      inputVarIsLink:
        description: 如果来自变量, 那么记录是来自 其它链还是其它Node
        type: boolean
      isParam:
        description: 位置, 1.入参 2.出参
        type: integer
      linkInstanceID:
        description: 所属 链 instance id
        type: integer
      name:
        description: 字段名
        type: string
      outputConst:
        description: 如果是确定值, 那么记录这个值
        items:
          type: integer
        type: array
      outputType:
        description: 返回值类型 (如果返回变量, 那么不需要记录, 到时候直接用就可以)
        type: integer
      type:
        description: 字段类型
        type: integer
    type: object
  model.NodeDefineInfo:
    properties:
      belongLinkDefineID:
        description: 所属 链 ID
        type: integer
      comment:
        description: 注释
        type: string
      hasInstance:
        type: boolean
      id:
        type: integer
      name:
        description: 方法名
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.NodeParamDefine'
        type: array
      preNodeDefineID:
        description: 前一个NodeID, 第一个写-1
        type: integer
    type: object
...
  model.NodeParamDefine:
    properties:
      comment:
        description: 字段注释
        type: string
      hasInstance:
        type: boolean
      id:
        type: integer
      isList:
        description: 是否为 列表
        type: boolean
      isParam:
        description: 位置, 1.入参 2.出参
        type: integer
      mapKeyParamID:
        description: 当ParamType 为 map 时
        type: integer
      mapValueParamID:
        type: integer
      name:
        description: 字段名
        type: string
      nodeDefineID:
        description: 所属节点id
        type: integer
      sort:
        description: 排序位置
        type: integer
      structID:
        description: 对应结构体ID, 当 FType 为 ParamTypeStruct 时 有用
        type: integer
      type:
        description: 字段类型
        type: integer
    type: object
...
  model.UpdateLinkDefineReq:
    properties:
      comment:
        description: 注释
        type: string
      hasInstance:
        type: boolean
      id:
        type: integer
      isShared:
        description: 是否共享, 不共享的话 只能当前工程可用
        type: boolean
      name:
        description: 链名称
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.LinkParamDefine'
        type: array
      projectDefineID:
        description: 所属工程ID
        type: integer
    type: object
  model.UpdateLinkDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 更新的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.UpdateLinkInstanceReq:
    properties:
      comment:
        description: 注释
        type: string
      defineID:
        description: 所用 链 定义 id
        type: integer
      id:
        type: integer
      isShared:
        description: 是否共享, 不共享的话 只能当前工程可用
        type: boolean
      name:
        description: 链名称
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.LinkParamInstance'
        type: array
      projectInstanceID:
        description: 所属工程ID
        type: integer
    type: object
  model.UpdateLinkInstanceResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 更新的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
  model.UpdateNodeDefineReq:
    properties:
      belongLinkDefineID:
        description: 所属 链 ID
        type: integer
      comment:
        description: 注释
        type: string
      hasInstance:
        type: boolean
      id:
        type: integer
      name:
        description: 方法名
        type: string
      paramList:
        items:
          $ref: '#/definitions/model.NodeParamDefine'
        type: array
      preNodeDefineID:
        description: 前一个NodeID, 第一个写-1
        type: integer
    type: object
  model.UpdateNodeDefineResp:
    properties:
      code:
        description: 结果码
        type: integer
      id:
        description: 更新的ID
        type: integer
      msg:
        description: 提示信息
        type: string
    type: object
...
info:
  contact:
    email: guoyueyang@126.com
    name: author:guoyueyang
    url: http://www.amazingsaltedfish.com
  description: this is [amazing salted fish] swagger api page.
  license:
    name: License:MIT
    url: https://spdx.org/licenses/MIT.html
  termsOfService: http://www.amazingsaltedfish.com
  title: Amazing Salted Fish API
  version: 0.0.1
paths:
  /link/define/:
    post:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.CreateLinkDefineReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateLinkDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 新建一个结构信息
      tags:
      - link定义信息
      - 新建
      - 定义
    put:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.UpdateLinkDefineReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdateLinkDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 更新一个结构信息
      tags:
      - link定义信息
      - 更新
      - 定义
  /link/define/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: project的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeleteLinkDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 删除一个结构信息
      tags:
      - link定义信息
      - 删除
      - 定义
  /link/define/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetLinkDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取一个结构信息
      tags:
      - link定义信息
      - 获取
      - 定义
  /link/define/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页信息, 页数从1开始
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 分页信息, 每页最大500
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetLinkDefineListResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取结构信息列表
      tags:
      - link定义信息
      - 获取
      - 列表
      - 定义
  /link/instance/:
    post:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.CreateLinkInstanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateLinkInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 新建一个结构信息
      tags:
      - link实体信息
      - 新建
      - 实体
    put:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.UpdateLinkInstanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdateLinkInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 更新一个结构信息
      tags:
      - link实体信息
      - 更新
      - 实体
  /link/instance/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: project的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeleteLinkInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 删除一个结构信息
      tags:
      - link实体信息
      - 删除
      - 实体
  /link/instance/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetLinkInstanceResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取一个结构信息
      tags:
      - link实体信息
      - 获取
      - 实体
  /link/instance/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页信息, 页数从1开始
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 分页信息, 每页最大500
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetLinkInstanceListResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取结构信息列表
      tags:
      - link实体信息
      - 获取
      - 列表
      - 实体
  /node/define/:
    post:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.CreateNodeDefineReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateNodeDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 新建一个结构信息
      tags:
      - node定义信息
      - 新建
      - 定义
    put:
      consumes:
      - application/json
      parameters:
      - description: 结构体信息
        in: body
        name: projectInfo
        required: true
        schema:
          $ref: '#/definitions/model.UpdateNodeDefineReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UpdateNodeDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 更新一个结构信息
      tags:
      - node定义信息
      - 更新
      - 定义
  /node/define/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: project的ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DeleteNodeDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 删除一个结构信息
      tags:
      - node定义信息
      - 删除
      - 定义
  /node/define/detail/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetNodeDefineResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取一个结构信息
      tags:
      - node定义信息
      - 获取
      - 定义
  /node/define/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页信息, 页数从1开始
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 分页信息, 每页最大500
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetNodeDefineListResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.BaseResp'
      summary: 获取结构信息列表
      tags:
      - node定义信息
      - 获取
      - 列表
      - 定义
...
swagger: "2.0"